fix(raster): preserve power and rotated-grid identity

+++ src/core/job/compile-job-raster-stream.test.ts
@@ -88,7 +88,7 @@ function materializedReference(
   mask: SceneObject | null,
   device: DeviceProfile,
   algorithm: DitherAlgorithm,
-): Uint16Array {
+): Float64Array {
   const resampled = resampleLumaNearest(
     { luma: sourceLuma(), width: SOURCE_W, height: SOURCE_H },
     TARGET_W,
